Add initial implementation of creating and deleting Group
Creating a new Group should scroll it to the middle of the Sidebar and put it into rename mode. Ive tried with no
success to have SwiftUI scroll the newly-created Group into view and some of the code that enables this is left
commented here. Renaming is not possible yet. Deleting a Group should present a prompt, but I haven't figured
out how SwiftUI handles that yet.

The Swift compiler is having a really hard time crunching through the SwiftUI Views at this point and I'm not sure
what I should be doing to simplify things for it. deCONZClientModel has been extracted into its own file and some
actions pulled out to functions to at least keep compilation finishing - but it's taking quite a while for such a small
project.