Complete 'add', 'rename', and 'delete' implementation for Groups
In order to work around a possible bug in ScrollViewReader, this commit introduces the ability
to take a "snapshot" of the Sidebar's items without including any non-disclosed nodes. This let's
ScrollViewReader correctly scroll to items. The "snapshot" is removed after the scrolling, and the
original list restored. This now allows for the desired "create new Group" experience, which is now
fully implemented along support for renaming and deleting Groups.

There remains an issue related to tracking of which DisclosureGroups are expanded - the information
will currently be lost when refreshing the list of Groups and Scenes over the deCONZ REST API.