Add an interface for adding Lights to and removing lights from Scenes
Lights aren't actually added to or removed from Scenes in deCONZ, but to or from Groups. All Scenes
in a Group have access to the same Lights. The interface doesn't make this distinction clear and might
lead users to unexpected situations. Maybe this could be further clarified in the sheet that will be
presented with the list of Lights that can be added.