Add initial support for adding/removing Lights to/from Groups
Supporting this without affecting the current State Attributes of Lights turned out to be quite complicated.
Keeping the list of Lights in a Group in sync with the list of Lights in each of the Group's Scenes requires
lots of calls to the deCONZ REST API, which turn into even more requests at the Zigbee network level.
There's currently no throttling or error handling, so trying to modify Groups with many Lights or many Scenes
will likely cause errors and leave the Group/Scenes in an inconsistent state. This feature needs lots more
testing and should be considered very unstable right now.