Separate Lights in Groups from Lights in Scenes
The previous implementation attempted to synchronize Lights across Groups and their Scenes and wasn't
very successful, Instead, I'm now trying to get deCONZ's REST API modified to allow for the Lights in a
Group and the Lights in their Scenes to be different. For a Light to be in a Group's Scene it will need to
belong to the Group, but it will be possible for a Scene to have only a subset of the Group's Lights.