Allow modifying Scene Lights
Modifying Lights in Scenes requires a version of the deCONZ REST API that has been built with
my patches to allow adding Lights to a Scene that didn't previously belong to it, so long as they're
part of the same Group and removing Lights from a Scene by supplying an empty body as the Light's
state.

There's a few TODOs in the code - those can be tackled if and when the patches to deCONZ are
pulled into the mainline.