Refactor deCONZ REST API data model
The purpose of this rather heavy refactoring was to reduce the number of Optional checks
that had to be performed when accessing data in the models. While successful, it has come
at the expense of complicating the data model and requiring conversions between private and
public data types. On top of that, it doesn't seem like the call sites were much improved.