Merge attempts at features back into mainline
This very messy commit is an attempt at unifying three different streams
the project has branched off in the last year:

* Migration to Swift 6 and Strict Concurrency
* SwiftUI data flow rearchitecture
* deCONZ client and model rearchitecture

All streams where started and abandoned multiple times, and some parts
of the work have been done as indepnedent projects. The commit history
for most of them was mangled or lost in a combination of workstation and
storage migrations.

Merging the three streams leaves the tool in a “builds and runs” state,
but most features are incomplete and not in a working state. The
codebase is messy and littered with commented-out code that needs to be
pruned, but at least now it’s in a state where all this can happen under
version control.