Fix ‘esc’ and ‘return’ actions when renaming Sidebar items
Pressing escape while editing a Sidebar item was removing focus from
the Sidebar. This makes it undo the change, close the editor, but leaves
the Sidebar item selected.

Pressing return while editing a Sidebar item was not closing the editor.
Checks for Sidebar item in question already being in rename mode in the
‘onKeyPress’ modifier associated to the list were missing and have now
been added.