Fix insertion of new SidebarItems into Sidebar
A missing check when creating ‘scene’ types blocked the sorting of new
SidebarItems after being renamed. The items can now be inserted and
sorted. The necessary calls to the REST API are still missing and need
to be selected between creating a new group/scene or renaming an
existing group/scene.