Add CRUD operations for Groups and Scenes
Groups and Scenes can now be created, renaned, and deleted from the
interface. The commands are forwarded to the REST API, so the operations
happen there as well.

There is minimal error handling - this is focused on just making the
operations work in order to bring the ‘unification’ effort to feature
parity with the previous version of the application.