Add functionality to save scene presets
The scene editor’s content can now be saved as a preset and persisted
in the application’s container as a .json file. The functionality is
accessed through a button in the application’s toolbar that presents
a popover where the preset can be named and saved. If a preset with the
same name already exists, it is overwritten with the new values.

In order to access the state text as JSON, the scene editor’s text
storage has been refactored out to the Window model.