Remove ‘colormode’ from LightState objects
Not all parts of the REST API reference ‘colormode’ as part of a light’s
state. To provide consistency, the model layer will drop all references
to it everywhere. It’s already not in use, but some serialized scene
and preview data still had them.

The model layer could infer the color mode by checking if the LightState
object has values in their ‘xy’ or ‘ct’ properties.