Fix light list sizing and scrolling
Attempt to improve the height of rows in the list. These values
seem nicer when using short and long lists of lights.

In order to prevent scrolling of the list when there are few lights,
the bounce behavior of the scroll view is set to ‘.basedOnSize’.
Unfortunately, this also prevents “pulling” the list up or down with
the trackpad when the list is long enough to scroll.