Display fixture and bulb images in light lists
deCONZ provides a “modelID” property that can be used to identify
a light’s model. References to it have been propagated from RESTLight to
Light and then used in LightItem to select an image that fits the light
model.

For situations where the bulbs originally shipped in fixtures have been
replaced, the light’s name is sometimes also used to determine the
image to use. This requires the light’s name to contain certain
identifiers that match the current naming conventions I use with my
lights.