Create data models from Hue mobila app scenes
The dragged and dropped .json files get processed into a data model that
represents the structure in the JSON and are then converted to a data
model that closely resembles the one used by Presets in SceneManager.
Ideally, this will be the same data model in the future, but the one
used in SceneManager doesn’t support Dynamic Scenes yet.