Show scene colors in PresetItemView
For Dynamic Scenes, the color palette will be displayed as a collection
of colored circles in the bottom of the view. Regular scenes will also
show a circle with their color. The image representing the scene has
been made more “stylish” by enlarging it, moving it slightly off-scene,
and giving it some translucency.

The view’s background continues to be the color of the first (or only)
color in the list. Changing the background to an image or a gradient
of sorts could make it look better.