Fix visual overflow with addtional colors
Presets with more than 6 colors (of the 9 nine allowed) were overflowing
the inspector with their color circles. The spacing between the circles
is adjusted depending on the number of colors to have the all fit.

The shadows have been slightly softened, as they were too hard on the
more compacted circles.