Refactor Presets support of colors and effects
This massive refactor of Presets aims to provide a type-safe definition
of the preset’s underlying state, which is built off of JSON strings.

The many structs introduced are there to represent, in a type-safe way,
the parts of scene states that are supported - both for regular scenes
and for dynamic scenes.

The new models will also simplify the way dynamic presets can be created
within the app and allow for also displaying (tinted) effect circles
in PresetItemViews that have effects (both not supported yet).