Improve handling of Dynamic Scene attributes
Better handling of attributes that are not part of Dynamic Scenes, but
are used by the app when applying them to recall states.

‘scene_apply’ - which indicates to the app the order in which to apply
dynamic states to the recall state, is injected as ‘sequence’ in cases
where it was missing.

‘transitiontime’ - is present on the dynamic definitions, even though it
is not used on dynamic scenes. It’s simply passed on to the recall
state.