Fix redraws in AddView
The view's height was inadvertently being set while the view's task
was fetching the LightItems to display. Because the view height is
dependent on the number of LightItems, this was causing view redraw
loops.

The LightItems are now computed before displaying AddView and passed
in as a parameter. A benefit of doing it this way is that AddView is
immediately populated with LightItems when its sheet pulls up.