Refactor error propagation
RESTClient, RESTModel, and WindowItem will chain throwing any resulting
errors back to their callers, which in most cases are SwiftUI Views.

An extension on Task has been introduced, allowing a 'catch' closure
to be passed in where Errors can be handled. This commit prepares the
codebase for presenting those errors as Alerts to the user and marks
the call sites with #warning for obvious and quick identification.