Add warning indicator to window title
There is no provision in SwiftUI to position a toolbar item after the
window's title. In order to do this, the window's title is blanked out
and a replacement has been constructed as a toolbar item. It mimics the
original as close as possible, and allows the indicator to follow it.

The warning indicator should be visible when certain REST API calls
or encoding/decoding operations fail. Calls to the REST API that create,
delete, or rename should show an alert dialog instead. Calls that
create, delete, or rename Presets should likewise should alert dialogs.

The display state of the warning indicator will have to be controlled by
the do/catch blocks that can cause it, as they are also the ones best
suited to reset it when called.