Add activity tracker for RESTClient
RESTClient calls to the deCONZ REST API have been extended to record
their request, response, and outcome to an optional activity tracker.

RESTModel can now own a RESTActivity which will keep a record of the
operations performed in a session. The warning indicator introduced
previously can be used to let the user know when an operation has
failed and provide some details as to why.