Add error handling with Alert view presentation
Introduce an AlertView that displays errors as a popover from the
warning indicator introduced previously. Parts of the UI code that
catch errors will set a flag on WindowItem to indicate that the warning
indicator should be displayed. The flag is expected to be cleared when
a new operation is started.

The JSON editors for scene descriptions are the first to receive error
handling. Potential errors coming from the REST API are still just
forwarded for now, but DecodingError arising from JSON parsing is
handled specifically for the common cases. This is the area where user
errors are most likely to occur, making error handling here very
important.