Handle errors in Presets
The order of operations has been refactored so there’s no need to undo
changes to the model if operations at the filesystem level fail. Changes
happen first at the filesystem level and are updated in the model only
if they were successful.

Renaming a preset is the exception. Because the rename happens directly
on the model, a variable shadowing the PresetItem’s name has been added
to PresetItem. It is updated to the PresetItem’s current name when
isRenaming is set to ‘true’. Should the rename fail, ‘restoreName()’
can be called to copy the previous name back into the PresetItem. This
is preferred over reading the value from the filesystem again.