Add functionality to delete Dynamic Scenes
Dynamic Scenes can now be removed from Scenes that have them. The recall
state of the scene is not affected, but all dynamic actions are removed
from the scene.

Dynamic Scenes are removed by right-clicking on their Scene in the
sidebar and selecting 'Delete Dynamic Scene' from the context menu. The
option is only visible for Scenes that have a Dynamic Scene.