Add initial implementation of light configuration UI
The first attempt at a UI for setting light configuration comes with a
couple of issues. The most annoying one is broken focus navigation with
the keyboard.

SwiftUI has no provision for moving the focus from row to row in
TableViews. Focus navigation works as expected when using a LazyVStack
instead, but only cycles through visible views instead of scrolling the
next/previous row into view and continuing. This can be solved later.

The Views in the implementation make heavy use of computed properties
that return custom Bindings to address only the filtered lights via the
“global” input controls at the top of the UI.