Add static progress view for operations
The static progress view serves as a test for the 300ms each REST API
call will have to wait for. The REST API provides responses when the
HTTP request is processed, not when the Zigbee request is processed.
The delay is intended to give the Zigbee network enough time to
propagate the message while not adding too much time to operations.