Implement mass light startup configuration
The interface for configuring the startup values for lights is now
hooked up to the REST API and working.

RESTClient has a new method to set the startup configuration values in
lights and underwent refactoring to allow multiple errors to be present
in a response, which was noticed as a possibility while debugging.

Handling errors with Warning indicator has been extended to allow for
multiple errors coming from the REST API.

The data models related to light configuration in RESTLight also
required refactoring to correctly map enum values to JSON output.