Add functionality to delete Phoscon keys
Instead of remembering and reusing its last key, Phoscon creates a new
REST API key every time a user logs in. It does so using the "Basic
Authentication" method, so the gateway does not need to be unlocked for
it to create the key. While likely harmless, it's annoying.

It's now possible to delete these keys via the app's "deCONZ" menu, by
selecting "Delete Phoscon Keys…". The implementation could be cleaned
up and extracted out of 'SceneManagerApp'. A progress bar showing the
keys being deleted would also provided useful feedback to the user.