Refactor app and WindowItem to allow multiple windows
'ContentView' now owns the data models that represent the contents of
a window, instead of 'SceneManagerApp'. This allows SwiftUI to create an
instance of the models per Window in a WindowGroup, enabling the app
to have multiple windows. The app still only allows a single connection
to a REST API that is shared by all windows.

The Menu Bar items have been refactored to affect the last window to
have had focus. This is tracked using SwiftUI's 'focusedSceneValue'
which is applied to 'ContentView' and holds the WindowItem for the last
focused window.