Refactor Presets for multiple window support
The refactor to Presets needed to be extensive and touched various parts
of the code that needed to be updated together to avoid breaking -
hinting that the way Presets are architectured could use a refactor.

JSON.swift has been pulled out of the deCONZ library. While not tied to
this refactor, it does affect the imports Presets and the new
PresetsModel need to make.

PresetsModel is a singleton, very similar to RESTModel that centralizes
operations to presets and takes care of persisting and loading them
to and from the filesystem.

PresetsModel holds a cache of available presets that it shares over a
Combine Publisher with Preset models. The cache is made up of classes -
PresetItemGroup and PresetItem. It is important that each instance of
'Presets' (one per Window) holds its own copy of those classes to be
able to modify them independently - which is what the code in the
Publisher's 'sink' does.