Use Swift 6 to build the app
Swift 5 was still being used to build the app and a few concurrency
warnings had not been addressed. A couple of classes related to Presets
that do work on UI items have been marked as constrained to MainActor.

The biggest change is having do/catch errors directly within tasks - the
extension that had been created previously was susceptible to race
conditions when passing the error handling closure around.