Adjust PresetItemView width
Tighten up the spacing for PresetItemViews that have more that 4
effect/colors defined in them. The previous spacing looked fine until
scrollbars are visible when setting macOS System Settings -> Appearance
-> Show scroll bars to “Automatically based on mouse or trackpad”
because a mouse is present.

It would be nicer to find a way to determine the spacing for the circles
based on the width available, instead of a constant derived from visual
experiments.