Import color pickers from side project
The last bits of code to have spun off as a side project are now
integrated into the app. These are two colors pickers that can be used
to visually select colors or color temperatures by moving "droppers"
around a color wheel.

The implementations of ColorGradientPicker and ColorTemperaturePicker
share lots of code. It could be worthwhile to invest some time in the
future in centralizing the shared bits in a way that works nicely with
Swift structs.