Import color picker updates from side project
ColorGradientPicker has undergone many changes to make its color lookup
and interpolation based on HSB (used to be RGA). This makes it possible
to reliable and deterministically look up colors and locations of colors
within the wheel. Practically, this means that two color pickers with
different (but ideally some shared) color spectrums in a TabView can
share picker droppers - the droppers will correctly locate themselves in
the pickers as the tab selection changes.

Errors in the pickers are now thrown, instead of crashing the picker.

A very simple zIndex ordering for multiple pickers has been added. In
short, the last selected dropper is always ordered at the top. For the
most part, this works well - but does not preserve zIndex ordering
across View redraws.