Update UI to macOS Tahoe
Various visual tweaks to make the application feel more at home in
macOS 26.

Buttons for applying scene definitions to scenes or lights have been
moved into the toolbar, along with the code for handling the actions,
which should be refactored later on.

The Lights list has had its title removed and the bottom bar has
received a glass effect - as has the Sidebar’s bottom bar.

Presets has also received a few tweaks and had a few workarounds removed
as now most of it seems to work well with SwiftUI. Its filter field did
not receive a glass effect yet as it didn’t work with just the provided
view modifiers.