Move Presets filter field to toolbar
Moving the input field used for filtering the list of Presets to the
toolbar, which is where macOS 26 prefers search bars to be. This makes
the application’s interface closer to the system applications.

Previously, the filter bar used a filter icon instead of the default
magnifying glass. After moving the filter bar to the toolbar, the
override for the icon no longer works.